requirements.txt
paho-mqtt
PyYAML
pyserial